% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RI_iarm_person_estimates}
\alias{RI_iarm_person_estimates}
\title{Temporary fix for upstream bug in \code{iarm::person_estimates()}}
\usage{
RI_iarm_person_estimates(object, properties = F, allperson = F)
}
\arguments{
\item{object}{Output from PCM() or RM()}

\item{properties}{All properties or not}

\item{allperson}{All respondents or not}
}
\description{
To get \code{RIscoreSE()} working properly for cases with theta range up til
c(-10,10).
}
\details{
code from package \code{iarm}
\url{https://github.com/cran/iarm/blob/master/R/Person-Fit.R}
}
