% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RI_iarm_persons_mle}
\alias{RI_iarm_persons_mle}
\title{Temporary fix for upstream bug in \code{iarm::person_estimates()}}
\usage{
RI_iarm_persons_mle(
  respm,
  thresh,
  model = c("RM", "PCM"),
  theta = rep(0, dim(respm)[1]),
  type = c("MLE", "WLE"),
  extreme = TRUE,
  maxit = 20,
  maxdelta = 3,
  tol = 0.0001,
  maxval = 9
)
}
\arguments{
\item{respm}{temp}

\item{thresh}{temp}

\item{model}{temp}

\item{theta}{temp}

\item{type}{temp}

\item{extreme}{temp}

\item{maxit}{temp}

\item{maxdelta}{temp}

\item{tol}{temp}

\item{maxval}{temp}
}
\description{
To get \code{RIscoreSE()} working properly for cases with theta range up til
c(-10,10).
}
\details{
code from package \code{iarm}
\url{https://github.com/cran/iarm/blob/master/R/Person-Fit.R}
}
