% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIbardiv}
\alias{RIbardiv}
\title{Create a stacked diverging bar graph to show response distribution}
\usage{
RIbardiv(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Create a stacked diverging bar graph to show response distribution
}
