% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIbarplot}
\alias{RIbarplot}
\title{Create individual bar plots for all items.}
\usage{
RIbarplot(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Create individual bar plots for all items.
}
