% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIbarstack}
\alias{RIbarstack}
\title{Create a stacked bar graph to show response distribution}
\usage{
RIbarstack(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Create a stacked bar graph to show response distribution
}
