% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIcheckdata.R
\name{RIcheckdata}
\alias{RIcheckdata}
\title{Check dataframe}
\usage{
RIcheckdata(data, n = 3)
}
\arguments{
\item{data}{Font family for all plot text}

\item{n}{Lowest number of responses in a cell}
}
\value{
Whether there are issues with the data or not (TRUE or FALSE)
}
\description{
This functions checks for items with low number of responses in any category.
}
\details{
For now, low number is set to < 3.
}
