% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIcolorlistitems}
\alias{RIcolorlistitems}
\title{Show items based on itemlabels file, with coloring options}
\usage{
RIcolorlistitems(items, color)
}
\arguments{
\item{items}{vector of row numbers for items to colorize background}

\item{color}{color of background ("lightblue" is default)}
}
\value{
A table with items used in dataframe
}
\description{
Requires a dataframe with two columns, labeled "itemnr" and "item",
containing information on the item numbers (qN) and item content.
This dataframe has to be labeled "itemlabels".
}
\details{
Input a vector of item rows, i.e c(1,3,5) to colorize items 1, 3 and 5.
Optionally choose which background color will be used. "Lightblue" is the
default. Text will be black, so choose a light color which gives good
contrast for readability.
}
