% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdemographics}
\alias{RIdemographics}
\title{Create table for demographic variables}
\usage{
RIdemographics(dif.var, label)
}
\arguments{
\item{dif.var}{A vector with a demographic variable}

\item{label}{What the variable represents (sex/age/etc)}
}
\description{
Input should be a vector with a demographic variable such as gender or age.
}
