% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifFigTime}
\alias{RIdifFigTime}
\title{Create a DIF line graph over time, showing groups' PCM item locations}
\usage{
RIdifFigTime(dfin, dif.var)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}
}
\description{
This function is specifically intended for examining DIF over time, to
create a figure that follows each item location on the y axis
with time on the x axis.
}
