% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifFigureRM}
\alias{RIdifFigureRM}
\title{Create a DIF line graph, showing groups' RM item locations}
\usage{
RIdifFigureRM(dfin, dif.var)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}
}
\description{
Create a DIF line graph, showing groups' RM item locations
}
