% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifTable2}
\alias{RIdifTable2}
\title{DIF PCM interaction analysis}
\usage{
RIdifTable2(dfin, dif.var1, dif.var2, cutoff = 0.5)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{cutoff}{Cutoff in item location logit difference for table highlighting}

\item{dif.var}{DIF variable}
}
\description{
Makes use of the psychotree package. This function is for interaction between
two DIF variables
}
\details{
DIF variables need to be vectors with the same length as the number of rows
in the dataset.

sample usage: \code{RIdifTable2(df, dif.age, dif.gender)}
}
