% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifTableLR}
\alias{RIdifTableLR}
\title{DIF PCM analysis with table output for item locations}
\usage{
RIdifTableLR(
  dfin,
  dif.var,
  model = "PCM",
  sort = FALSE,
  fontfamily = "sans-serif",
  cutoff = 0.5
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}

\item{model}{Defaults to "PCM", optional "RM" under development}

\item{sort}{Set to TRUE to sort the table based on DIF size}

\item{fontfamily}{Set table font}

\item{cutoff}{Cutoff in item location logit difference for table highlighting}
}
\description{
Makes use of the eRm package function \code{LRtest()}. Outputs a table with item
average locations, group differences, and standard errors.
}
\details{
DIF variables need to be factors with the same length as the number of rows
in the dataset.

sample usage: RIdifTableE(df, dif.age)
}
