% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifTableRM}
\alias{RIdifTableRM}
\title{DIF analysis dichotomous - requires having set up dif.variables}
\usage{
RIdifTableRM(dfin, dif.var)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}
}
\description{
Makes use of the psychotree package, which also allows for interactions
between DIF variables, which is not implemented in this function (yet).
}
\details{
DIF variables need to be vectors with the same length as the number of rows
in the dataset.

sample usage: RIdifTable(df.omit.na, dif.age)
}
