% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifThreshTblLR}
\alias{RIdifThreshTblLR}
\title{DIF PCM analysis with table output for item thresholds}
\usage{
RIdifThreshTblLR(dfin, dif.var, fontfamily = "sans-serif", cutoff = 0.5)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}

\item{fontfamily}{Set table font}

\item{cutoff}{Cutoff in item location logit difference for table highlighting}
}
\description{
NOTE: currently only works with two DIF groups
}
\details{
Makes use of the eRm package function \code{LRtest()}. Outputs a table with item
average locations, group differences, and standard errors.

DIF variables need to be factors with the same length as the number of rows
in the dataset.

sample usage: RIdifTableThreshE(df, dif.age, fontfamily = "Arial")
}
