% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIestThetas}
\alias{RIestThetas}
\title{Person location estimation}
\usage{
RIestThetas(data, model = "PCM", method = "WLE")
}
\arguments{
\item{model}{Rasch model to use ("RM" for dichotomous data)}

\item{method}{Estimation method (defaults to "WLE")}

\item{dfin}{Dataframe with response data only (no demographics etc), items as columns}

\item{theta_range}{Range of theta (person location) values}
}
\description{
Outputs a dataframe of person locations and measurement error (SEM) for each person
}
\details{
Uses \code{iarm::person_estimates()} to estimate person locations
(thetas) for a dataframe with item data as columns and persons as rows.

Defaults to use WLE estimation (lower bias than MLE, see Warm, 1989) and PCM.

Note: If you want to use a pre-specified set of item parameters, please use
\code{RIestThetasOLD()}.
}
