% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIestThetasOLD}
\alias{RIestThetasOLD}
\title{Person location estimation (old version)}
\usage{
RIestThetasOLD(
  dfin,
  itemParams,
  model = "PCM",
  method = "WL",
  theta_range = c(-7, 7)
)
}
\arguments{
\item{dfin}{Dataframe with response data only (no demographics etc), items as columns}

\item{itemParams}{Optional item (threshold) location matrix}

\item{model}{Rasch model to use (currently only default PCM works)}

\item{method}{Estimation method (defaults to "WL")}

\item{theta_range}{Range of theta (person location) values}
}
\description{
NOTE: Does not work with dichotomous data
}
\details{
Outputs a vector of person locations, one for each row in the dataframe.

Uses thetaEst function from catR package to estimate person locations
(thetas) for a dataframe with item data as columns and persons as rows.
Defaults to use WL estimation (lower bias than ML) and PCM.
See ?thetaEst for options available.

A version for multi-core processing is available as \code{RIestThetasOLD2()}.
}
