% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIgetfit}
\alias{RIgetfit}
\title{Get simulation based cutoff values for item fit values}
\usage{
RIgetfit(data, iterations = 100, cpu = 4, method = "PCM")
}
\arguments{
\item{data}{Dataframe with response data}

\item{iterations}{Number of simulation iterations}

\item{cpu}{Number of CPU cores to use}

\item{method}{PCM or RM, currently only PCM works}
}
\description{
This function uses your response data to simulate datasets that fit the
Rasch model to find a credible range of item fit values. The function
outputs an object that is strongly recommended to save to a file, since it
takes some time to run this function when using many iterations/simulations.
}
\details{
The output is a list object, which can in turn be used with three different
functions. Most importantly, you can use it with \code{RIitemfitPCM()} to set
cutoff values for both MSQ and ZSTD based on your sample size and item
parameters. This uses the lowest and highest value from \emph{all} items as cutoff
for all items.

If you want to see a table with the individual item fit upper/lower cutoffs,
the function \code{RIgetfitTable()} should be used with the output from this
function as input.

The function \code{RIgetfitPlot()} uses the package \code{ggdist} to plot the
distribution of fit values from the simulation results.
}
