% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIgetfitLoHi}
\alias{RIgetfitLoHi}
\title{Convenience function to get item fit cutoff values used in table}
\usage{
RIgetfitLoHi(gf)
}
\arguments{
\item{gf}{Output object from \code{RIgetfit()}}
}
\description{
Uses the output from \code{RIgetfit()} as input.
}
