% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIgetfitPlot}
\alias{RIgetfitPlot}
\title{Creates a plot with distribution of simulation based item fit values}
\usage{
RIgetfitPlot(simcut)
}
\arguments{
\item{simcut}{Output object from \code{RIgetfit()}}
}
\description{
Uses the output from \code{RIgetfit()} as input. Uses \code{median_qi}
and \code{.width = c(.66,.99)} with \code{ggdist::stat_dotsinterval()}.
}
