% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIheatmap}
\alias{RIheatmap}
\title{Create a Guttman-like tileplot/heatmap}
\usage{
RIheatmap(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Sorts items (y axis) and persons (x axis) according to respective
total score and displays them with a color gradient based on responses
}
