% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIinfitLoc}
\alias{RIinfitLoc}
\title{Plot with infit ZSTD and item location
ZSTD is sample size sensitive, see "RIitemfitPCM"
for options}
\usage{
RIinfitLoc(dfin, samplesize, nsamples)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{samplesize}{Desired sample size in multisampling (recommended 250-500)}

\item{nsamples}{Desired number of samples (recommended range 10-30)}
}
\description{
Plot with infit ZSTD and item location
ZSTD is sample size sensitive, see "RIitemfitPCM"
for options
}
