% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemfitPCM}
\alias{RIitemfitPCM}
\title{Create table with Rasch item fit values for each item.}
\usage{
RIitemfitPCM(dfin, jz, yz)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{jz}{Desired sample size in multisampling}

\item{yz}{Desired number of samples (recommended range 10-50)}
}
\description{
ZSTD is inflated with large samples (N > 500). Optional function to reduce
sample size to jz and run analysis using yz random samples to get average ZSTD
If you are using Quarto/Rmarkdown, "cache: yes" will be a useful chunk option to
speed things up. 50 samples seems to give stable output, but 10 is probably
sufficient for a quick look at the approximate ZSTD statistics. It is recommended
to use sample size 250-500, based on Hagell & Westergren, 2016.
}
