% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemfitPCM2}
\alias{RIitemfitPCM2}
\title{Create table with Rasch PCM model item fit values for each item.}
\usage{
RIitemfitPCM2(dfin, jz = 300, yz = 6, cpu = 4)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{jz}{Desired sample size in multisampling (recommended range 250-500)}

\item{yz}{Desired number of samples (recommended range 10-30)}
}
\description{
Special version of RIitemfitPCM that utilizes 4 CPU cores to improve
performance. Requires \code{library(doParallel)}. To find how many cores you
have on your computer, use \code{parallel::detectCores()}.
}
\details{
ZSTD is inflated with large samples (N > 500). Optional function to reduce
sample size to jz and run analysis using yz random samples to get average ZSTD
25 samples seems to give a stable output, but 10 is probably
sufficient for a reliable look at the approximate ZSTD statistics.
It is recommended #' to use sample size 250-500, based on
Hagell & Westergren, 2016.
}
