% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemHierarchy}
\alias{RIitemHierarchy}
\title{Create a figure showing items and thresholds (with 95\% CI)}
\usage{
RIitemHierarchy(dfin, numbers = TRUE)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{numbers}{Display text in figure with item threshold locations}
}
\description{
Items are sorted by item average location.
}
\details{
Only works for PCM models currently. For dichotomous data, use
\code{df.erm<-RM(data)} followed by \code{plotPImap(df.erm, sorted = T)}
}
