% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemparams}
\alias{RIitemparams}
\title{Item parameters summary}
\usage{
RIitemparams(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Displays a table with item threshold locations
and exports a CSV file ("itemParameters.csv") to your working
directory. This file can be used for estimating person thetas/scores.
}
