% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIlistitems}
\alias{RIlistitems}
\title{Show items based on itemlabels file}
\usage{
RIlistitems(dfin, pdf.out)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{pdf.out}{Set to TRUE to get PDF-compatible table (kableExtra)}
}
\value{
A table with items used in dataframe
}
\description{
Requires a dataframe with two columns, labeled "itemnr" and "item",
containing information on the item numbers (qN) and item content.
}
