% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RImissingP}
\alias{RImissingP}
\title{Creates a figure with item missing data descriptives for participants}
\usage{
RImissingP(data, itemStart, output, ...)
}
\arguments{
\item{data}{Dataframe/tibble to create table from}

\item{itemStart}{What your variable names start with, in quotes}

\item{output}{Optional dataframe with participants with missing data}

\item{...}{Options for \code{theme_rise()}}
}
\description{
Sample use: \code{RImissingP(df, itemStart = "PSS")}
}
\details{
If \code{itemStart} is missing, the whole dataframe will be used.
}
