% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIoutfitLoc}
\alias{RIoutfitLoc}
\title{Plot with outfit ZSTD and item location
ZSTD is sample size sensitive, see "RIitemfitPCM"}
\usage{
RIoutfitLoc(dfin, samplesize, nsamples)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{samplesize}{Desired sample size in multisampling}

\item{nsamples}{Desired number of samples (recommended range 10-50)}
}
\description{
Plot with outfit ZSTD and item location
ZSTD is sample size sensitive, see "RIitemfitPCM"
}
