% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpartgamDIF}
\alias{RIpartgamDIF}
\title{Partial gamma analysis of Differential Item Functioning}
\usage{
RIpartgamDIF(data, dif.var, output = "table")
}
\arguments{
\item{data}{A dataframe with response data}

\item{dif.var}{A vector with a DIF variable}

\item{output}{Defaults to a HTML table, optional "quarto" and "dataframe"}
}
\description{
A simple wrapper for \code{iarm::partgam_DIF()}. Filters results to only show
statistically significant relationships and sorts the table on the absolute
value of partial gamma.
}
\details{
Conditional highlighting in HTML table output set to partial gamma > +/-0.21.
}
