% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpartgamLD}
\alias{RIpartgamLD}
\title{Partial gamma analysis of local dependence}
\usage{
RIpartgamLD(data, output = "table")
}
\arguments{
\item{data}{A dataframe with response data}

\item{output}{Defaults to a HTML table, optional "quarto" and "dataframe"}
}
\description{
A simple wrapper for \code{iarm::partgam_LD()}. Filters results to only show
statistically significant relationships and sorts the table on the absolute
value of partial gamma.
}
\details{
Conditional highlighting in HTML table output set to partial gamma > +/-0.21.
}
