% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpboot}
\alias{RIpboot}
\title{Parametric bootstrapping outputting simulated datasets}
\usage{
RIpboot(data, iterations, cpu)
}
\arguments{
\item{data}{A dataframe with response data}

\item{iterations}{How many datasets to generate}

\item{cpu}{Number of CPU cores to use}
}
\description{
For generic use. Outputs datasets in a list object.
}
