% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpcm}
\alias{RIpcm}
\title{Running the Rasch PCM model using eRm, and
conducting a PCA of residuals to get eigenvalues.}
\usage{
RIpcm(dfin, no.table)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{no.table}{Set to TRUE to avoid output of table}
}
\description{
Running the Rasch PCM model using eRm, and
conducting a PCA of residuals to get eigenvalues.
}
