% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpcmPCA}
\alias{RIpcmPCA}
\title{Fits the Rasch PCM model using eRm, and
conducts a PCA of residuals to get eigenvalues using \code{psych::pca()}
and reports the top 5 values. Proportion of explained variance is calculated
using \code{stats::prcomp()}.}
\usage{
RIpcmPCA(dfin, output = "table", fontsize = 15)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{output}{Defaults to "table", optional "dataframe"}

\item{fontsize}{Set font size for table}
}
\description{
Note from \code{?psych::pca}:
The eigen vectors are rescaled by the sqrt of the eigen values to produce
the component loadings more typical in factor analysis.
}
