% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpfit}
\alias{RIpfit}
\title{Person fit}
\usage{
RIpfit(
  dfin,
  model = "PCM",
  pointsize = 2.5,
  alpha = 0.5,
  bins = 30,
  group,
  output = c("hist", "heatmap")
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{model}{Rasch model to use, "PCM" or "RM"}

\item{pointsize}{Size of datapoints for grouped view}

\item{alpha}{Transparency of points (0-1 where 1 = not transparent)}

\item{bins}{Number of bins for hexplot}

\item{group}{Optional grouping variable}

\item{output}{Can also be "rowid" for a vector of rownumbers}
}
\description{
Outputs a histogram of person fit ZSTD and a plot with person fit ZSTD and
person location/score. Defaults to output a histogram and a hex heatmap.
}
\details{
Optional grouped output with colorized points.

You can also get a vector with row numbers for persons with infit ZSTD
over/under +/- 2 by using \code{output = "rowid"}.
}
