% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIrawdist}
\alias{RIrawdist}
\title{Floor/ceiling effects based on raw data (ordinal scores)}
\usage{
RIrawdist(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\value{
A barplot with descriptives in footnote
}
\description{
Floor/ceiling effects based on raw data (ordinal scores)
}
