% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIrestscore}
\alias{RIrestscore}
\title{Item-restscore}
\usage{
RIrestscore(data, output = "table", sort, p.adj = "BH")
}
\arguments{
\item{data}{A dataframe with response data}

\item{output}{Defaults to a HTML table, optional "quarto" and "dataframe"}

\item{sort}{Optional sorting on absolute difference (descending)}

\item{p.adj}{See \code{?item_restscore()} for options (BH is default)}
}
\description{
A simple wrapper for \code{iarm::item_restscore()}
}
