% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RItileplot}
\alias{RItileplot}
\title{Create tile plot for all items, also showing the count of
responses in each response category for each item}
\usage{
RItileplot(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Create tile plot for all items, also showing the count of
responses in each response category for each item
}
