% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRaschSim.R
\name{SimPartialScore}
\alias{SimPartialScore}
\title{Simulate polytomous Rasch model data for a set of items}
\usage{
SimPartialScore(deltaslist, thetavec)
}
\arguments{
\item{deltaslist}{List object with item threshold locations}

\item{thetas}{Vector of person thetas/scores}
}
\description{
Documentation will be expanded on at a later point in time.
}
