% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRaschSim.R
\name{SimPolyItem}
\alias{SimPolyItem}
\title{Simulate polytomous Rasch model data for a single item}
\usage{
SimPolyItem(x, thetas)
}
\arguments{
\item{x}{Vector with item threshold locations}

\item{thetas}{Vector of person thetas/scores}
}
\description{
Documentation will be expanded on (and/or translated to English)
at a later point in time.
}
