% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{kbl_rise}
\alias{kbl_rise}
\title{A kableExtra function to simplify table code}
\usage{
kbl_rise(
  data,
  width = 65,
  fontsize = 14,
  fontfamily = "Arial",
  options = c("striped", "hover"),
  ...
)
}
\arguments{
\item{data}{Dataframe/tibble to create table from}

\item{width}{Width of table (0-100)}

\item{fontsize}{Font size}

\item{fontfamily}{Font family}

\item{...}{Passes options to kbl()}
}
\description{
A kableExtra function to simplify table code
}
