% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIallresp}
\alias{RIallresp}
\title{Create table with summarized responses across all items.}
\usage{
RIallresp(dfin, pdf.out)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{pdf.out}{Set to TRUE to get PDF-compatible table (kableExtra)}
}
\description{
Create table with summarized responses across all items.
}
