% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIbarstack}
\alias{RIbarstack}
\title{Create a stacked bar graph to show response distribution}
\usage{
RIbarstack(dfin, omit.na = F)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{omit.na}{Remove respondents with missing data (or not)}
}
\description{
Create a stacked bar graph to show response distribution
}
