% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIdifFigThresh}
\alias{RIdifFigThresh}
\title{Create a DIF line graph for item PCM thresholds}
\usage{
RIdifFigThresh(dfin, dif.var)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{dif.var}{DIF variable}
}
\description{
Produces a panel of linegraphs showing item thresholds over DIF nodes.
}
