% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIgetfit}
\alias{RIgetfit}
\title{Get simulation based cutoff values for item fit values}
\usage{
RIgetfit(data, iterations, cpu = 4, na.omit = TRUE)
}
\arguments{
\item{data}{Dataframe with response data}

\item{iterations}{Number of simulation iterations (use at least 1000)}

\item{cpu}{Number of CPU cores to use}

\item{na.omit}{Defaults to TRUE to produce conditional fit comparable values}
}
\description{
This function uses your response data to simulate datasets that fit the
Rasch model to find a credible range of item fit values. The function
outputs an object that is strongly recommended to save to an object, since it
takes some time to run this function when using many iterations/simulations.
}
\details{
The output is a list object, which can in turn be used with two different
functions. Most importantly, you can use it with \code{RIitemfit()} to get
conditional highlighting of cutoff values based on your sample size and
item parameters. Each item gets its own cutoff thresholds.

The function \code{RIgetfitPlot()} uses the package \code{ggdist} to plot the
distribution of fit values from the simulation results.

Uses multi-core processing. To find how many cores you have on your computer,
use \code{parallel::detectCores()}. Remember to keep 1-2 cores free.

Since version 0.2.4.2, the default is to only use complete cases in the
simulations, since this is what the conditional item fit function uses and
numbers should be more comparable using this method.
}
