% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIgetfitPlot}
\alias{RIgetfitPlot}
\title{Creates a plot with distribution of simulation based item fit values}
\usage{
RIgetfitPlot(gf, statistic)
}
\arguments{
\item{gf}{Output object from \code{RIgetfit()}}

\item{statistic}{Optionally one of infit_msq, outfit_msq, infit_zstd, outfit_msq}
}
\description{
Uses the output from \code{RIgetfit()} as input. Defaults to output all 4 plots.
}
\details{
Set a nicer x axis label by adding \code{+ xlab("Infit MSQ")}.
}
