% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemCats}
\alias{RIitemCats}
\title{Individual ICC plots.}
\usage{
RIitemCats(dfin, items)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{items}{A single item (e.g. "q4"), or a vector with multiple items (e.g. c("q4","q2")).}
}
\description{
Individual ICC plots.
}
