% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemfitPCM}
\alias{RIitemfitPCM}
\title{Create table with Rasch PCM model item fit values for each item.}
\usage{
RIitemfitPCM(
  dfin,
  samplesize,
  nsamples,
  zstd_min = -1.96,
  zstd_max = 1.96,
  msq_min = 0.7,
  msq_max = 1.3,
  fontsize = 15,
  fontfamily = "Lato",
  output = "table",
  tbl_width = 65,
  method = "conditional",
  simcut = FALSE,
  gf
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{samplesize}{Desired sample size in multisampling (recommended range 200-500)}

\item{nsamples}{Desired number of samples (recommended range 10-50)}

\item{zstd_min}{Lower cutoff level for ZSTD}

\item{zstd_max}{Upper cutoff level for ZSTD}

\item{msq_min}{Lower cutoff level for MSQ}

\item{msq_max}{Upper cutoff level for MSQ}

\item{fontsize}{Set fontsize for table}

\item{fontfamily}{Set font family for table}

\item{output}{Defaults to output a table. Optional "dataframe" or "quarto"}

\item{tbl_width}{Set table width in percent}

\item{method}{Defaults to "conditional". Optional "unconditional"}

\item{simcut}{Set to TRUE if you want to use simulation based cutoff values}

\item{gf}{The output object from \code{RIgetfit()} is needed when \code{simcut = TRUE}}
}
\description{
Defaults to using conditional estimates for MSQ values (Müller, 2020)
estimated using the \code{iarm} package. Use \code{method = "unconditional"} for the
"old" unconditional MSQ values (using \code{eRm}).
}
\details{
Since version 0.2.0 (2024-08-15), it is highly recommended to replace
rule-of-thumb cutoff values with simulation based cutoffs, see details in
\code{?RIgetfit()} for an easy way to get appropriate cutoff values.

ZSTD is inflated with large samples (N > 500). There is an optional function
to use a reduced sample size and run analysis using multiple random samples
to get the average ZSTD for each item over all runs.

If you are using Quarto, the YAML execute setting "cache: yes" will be a
useful chunk option to speed things up if you render often. 30-50 samples
seems to produce stable output, but 4-8 is probably sufficient for a quick
look at the approximate ZSTD statistics.
It is recommended to use sample size 200-500, based on
Hagell & Westergren (2016) & Müller (2020).
}
