% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemfitRM}
\alias{RIitemfitRM}
\title{Create table with Rasch dichotomous model item fit values for each item.}
\usage{
RIitemfitRM(
  dfin,
  samplesize,
  nsamples,
  zstd_min = -2,
  zstd_max = 2,
  msq_min = 0.7,
  msq_max = 1.3,
  fontsize = 15
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{samplesize}{Desired sample size in multisampling (recommended range 250-500)}

\item{nsamples}{Desired number of samples (recommended range 10-50)}

\item{zstd_min}{Lower cutoff level for ZSTD}

\item{zstd_max}{Upper cutoff level for ZSTD}

\item{msq_min}{Lower cutoff level for MSQ}

\item{msq_max}{Upper cutoff level for MSQ}

\item{fontsize}{Set font size for table}
}
\description{
ZSTD is inflated with large samples (N > 500). Optional function to reduce
sample size and run analysis using multiple random samples to get average ZSTD
If you are using Quarto/Rmarkdown, "cache: yes" will be a useful chunk option to
speed things up. 50 samples seems to give stable output, but 10 is probably
sufficient for a quick look at the approximate ZSTD statistics. It is recommended
to use sample size 250-500, based on Hagell & Westergren, 2016.
}
