% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemHierarchy}
\alias{RIitemHierarchy}
\title{Create a figure showing items and thresholds (with CI)}
\usage{
RIitemHierarchy(dfin, numbers = TRUE, sem_multiplier = 1.405)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{numbers}{Display text in figure with item threshold locations}

\item{sem_multiplier}{For confidence intervals displayed in figure}
}
\description{
Items are sorted by item average location. Confidence intervals are 84\% by
default to enable visual interpretation of statistically significant
differences (Payton et al., 2003). The CI can be changed using the
\code{sem_multiplier} option (ie. use 1.96 for 95\% CI).
}
\details{
Only works with partial credit models currently. For dichotomous data, use
\code{df.erm <- RM(data)} followed by \code{plotPImap(df.erm, sorted = T)}
}
