% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemHierarchy}
\alias{RIitemHierarchy}
\title{Create a figure showing items and thresholds (with 95\% CI)}
\usage{
RIitemHierarchy(dfin, ci = "95")
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{ci}{Show 95\% CI (default) around each threshold location, or "none"}
}
\description{
Items are sorted by item location.
}
\details{
Only works for PCM models currently.
}
