% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIitemparams}
\alias{RIitemparams}
\title{Item parameters summary}
\usage{
RIitemparams(
  dfin,
  fontsize = 15,
  output = "table",
  detail = "thresholds",
  filename = "item_params.csv",
  tbl_width = 90
)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{fontsize}{Option to set font size for table}

\item{output}{Defaults to "table, can be set to "dataframe" or "file"}

\item{detail}{Set to "all" to get more detailed summary output}

\item{filename}{Name of file to save output to}

\item{tbl_width}{Width of table}
}
\description{
Displays a table with item threshold locations. Can also output a dataframe or
a CSV file. Set \code{detail = "all"} to get more detailed output.
}
\details{
Currently only works with the Partial Credit Model (polytomous data).
}
