% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIlistItemsMargin}
\alias{RIlistItemsMargin}
\title{Create a table with the items used in a dataframe}
\usage{
RIlistItemsMargin(dfin, fontsize = 11)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{fontsize}{Defaults to 11, optimize if desired}
}
\description{
Depends on the \code{itemlabels} object.
}
\details{
Is intended to use with Quarto chunk option \code{column: margin}
}
