% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIloadLoc}
\alias{RIloadLoc}
\title{First residual contrast loadings vs item locations}
\usage{
RIloadLoc(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
First residual contrast loadings vs item locations
}
