% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIloadLoc}
\alias{RIloadLoc}
\title{Generates a plot showing the first residual contrast loadings based on a PCA
of Rasch model residuals vs item locations.}
\usage{
RIloadLoc(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\value{
A plot with item locations (y) and loadings (x)
}
\description{
Note. This function does not work with missing responses in the dataset.
You can temporarily remove respondents with missing data when running the
function, ie. \code{RIloadLoc(na.omit(df))}
}
