% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RImissing}
\alias{RImissing}
\title{Creates a figure with item missing data descriptives for items}
\usage{
RImissing(data, itemStart)
}
\arguments{
\item{data}{Dataframe/tibble to create table from}

\item{itemStart}{What your variable names start with, in quotes}
}
\description{
Sample use: \code{RImissing(df, itemStart = "PSS")}
}
\details{
If \code{itemStart} is missing, the whole dataframe will be used.
}
