% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpcmPCA}
\alias{RIpcmPCA}
\title{Fits the Rasch PCM model using eRm, and
conducts a PCA of residuals to get eigenvalues.}
\usage{
RIpcmPCA(dfin, no.table, fontsize = 15)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{no.table}{Set to TRUE to avoid output of table}

\item{fontsize}{Set font size}
}
\description{
Fits the Rasch PCM model using eRm, and
conducts a PCA of residuals to get eigenvalues.
}
