% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIpfit}
\alias{RIpfit}
\title{Person fit}
\usage{
RIpfit(dfin, model = "PCM")
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{model}{Rasch model to use, "PCM" or "RM"}
}
\description{
Outputs a histogram of person fit ZSTD and a plot with person fit ZSTD and
person location/score.
}
