% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIresidcorr}
\alias{RIresidcorr}
\title{Correlation matrix of Rasch residuals}
\usage{
RIresidcorr(data, cutoff, output = "table", ...)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{cutoff}{Relative value above the average of all item residual correlations}

\item{output}{Default HTML table, optional "quarto"}

\item{...}{Options sent to \code{kbl_rise()}}
}
\description{
Mandatory option to set relative cutoff-value over the average of all
item residual correlations. It is strongly recommended to use the function
\code{RIgetResidCor()} to retrieve an appropriate cutoff value for your data.
}
\details{
Christensen et al. (2017, p.181) write:
"under the null hypothesis, the average correlation of residuals is negative"
}
