% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIresidcorr}
\alias{RIresidcorr}
\title{Correlation matrix of Rasch residuals}
\usage{
RIresidcorr(dfin, cutoff, fontsize = 15, fontfamily = "Lato", tbl_width = 70)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{cutoff}{Relative value above the average of all item residual correlations}

\item{fontsize}{Set font size for table}

\item{fontfamily}{Set font family for table}

\item{tbl_width}{Set table width in percent}
}
\description{
Mandatory option to set relative cutoff-value over the average of all
item residual correlations. It is strongly recommended to use the function
\code{RIgetResidCor()} to retrieve an appropriate cutoff value for your data.
}
\details{
Christensen et al. (2017, p.181) write:
"under the null hypothesis, the average correlation of residuals is negative"

Note that negative values are currently not highlighted with red text, even
if they are above the cutoff value.
}
