% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIrespcats}
\alias{RIrespcats}
\title{Create a figure with ICC plots for all items (not working yet)}
\usage{
RIrespcats(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Create a figure with ICC plots for all items (not working yet)
}
