% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIscoreSE}
\alias{RIscoreSE}
\title{Raw sum score to logit score transformation table}
\usage{
RIscoreSE(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Displays a table with raw sum scores and their corresponding logit score
and logit standard error.
}
