% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RIscoreSE}
\alias{RIscoreSE}
\title{Raw sum score to logit score transformation table & figure}
\usage{
RIscoreSE(
  data,
  output = "table",
  point_size = 3,
  error_width = 0.5,
  error_multiplier = 1.96,
  ...
)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{output}{Options: "table" (default), "figure", or "dataframe"}

\item{point_size}{Point size for figure}

\item{error_width}{Width of error bar ends for figure}

\item{error_multiplier}{Range of error bars to multiply with SEM}

\item{...}{Options for \code{kbl_rise()} for table creation}
}
\description{
By default displays a table with raw sum scores and their corresponding logit score
and logit standard error. Depends on functions from package \code{iarm}.
}
\details{
Optional figure or dataframe output.

NOTE: the figure uses \code{coord_flip()}, take this into account if you wish to add theming.
}
