% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RItif}
\alias{RItif}
\title{Reliability of test}
\usage{
RItif(dfin, lo = -5, hi = 5, samplePSI = FALSE)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{lo}{Lower limit of x axis (default = -5)}

\item{hi}{Upper limit of x axis (default = 5)}

\item{samplePSI}{Adds information about sample characteristics}
}
\description{
Test information shows the reliability curve of the test (not the sample).
Use option \code{samplePSI = TRUE} to add graphical representation of the current
sample's theta mean/SD, mean/95\% CI test information (TIF), and Person
Separation Index (PSI) with 95\% CI.
}
