% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RItif}
\alias{RItif}
\title{Reliability, overall test information}
\usage{
RItif(dfin)
}
\arguments{
\item{dfin}{Dataframe with item data only}
}
\description{
Note: the eRm function below will provide individual item information curves
plotINFO(df.erm, type = "item", legpos = "topleft")
}
