% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RItif}
\alias{RItif}
\title{Reliability of test}
\usage{
RItif(dfin, lo = -5, hi = 5, samplePSI = FALSE)
}
\arguments{
\item{dfin}{Dataframe with item data only}

\item{lo}{Lower limit of x axis (default = -5)}

\item{hi}{Upper limit of x axis (default = 5)}

\item{samplePSI}{Adds information about sample characteristics}
}
\description{
Test information shows the reliability curve of the test (not the sample).
Use option \code{samplePSI = TRUE} to add graphical and written representation of
the current sample's theta mean/SD, test information (TIF) mean/SD , and
Person Separation Index (PSI). According to Wright & Stone (1999), PSI is
calculated as (SSD-MSE)/SSD, see \code{?eRm::SepRel} for details. According to Embretson &
Reise (2000), PSI = 1 - SEM^2, and TIF = 1/SEM^2, and the values reported in
this function are based on sample average SEM.
}
