% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{RItileplot}
\alias{RItileplot}
\title{Create tile plot for all items, also showing the count of
responses in each response category for each item}
\usage{
RItileplot(
  data,
  cutoff = 10,
  highlight = TRUE,
  percent = FALSE,
  text_color = "orange"
)
}
\arguments{
\item{data}{Dataframe with item data only}

\item{cutoff}{Conditional highlighting of text in cells with n below cutoff}

\item{highlight}{Defaults to TRUE. Set to FALSE to disable text highlighting}

\item{percent}{Set to TRUE to replace n with percentage of item responses}
}
\description{
Create tile plot for all items, also showing the count of
responses in each response category for each item
}
