% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISEkbmRasch.R
\name{theme_rise}
\alias{theme_rise}
\title{A simple ggplot theme for RISE formatting}
\usage{
theme_rise(
  fontfamily = "Lato",
  axissize = 13,
  titlesize = 15,
  margins = 12,
  axisface = "plain"
)
}
\arguments{
\item{fontfamily}{Font family for all plot text}

\item{axissize}{Font size for axis labels}

\item{titlesize}{Font size for plot title}

\item{margins}{Distance of axis labels to plot}

\item{axisface}{Set to "bold" if you want bold axis labels}
}
\value{
Add + theme_rise() to your ggplot or RIfunction that outputs a ggplot
}
\description{
Use is optional :)
}
\details{
See ?element_text for more details on available settings
}
